requirements.txt
agency-swarm>=0.1.0
ccxt>=4.1.13
pandas>=2.1.4
python-dotenv>=1.0.0
pydantic>=2.5.2
openai>=1.55.3,<2.0.0
docstring_parser==0.16
datamodel-code-generator==0.26.1
deepdiff==6.7.1
termcolor==2.4.0
rich>=13.7.0
jsonref==1.1.0
networkx>=3.0
numpy>=1.24.0
pandas-stubs>=1.2.0
psutil>=5.9.7
types-psutil>=5.9.0
asyncio>=3.4.3
uuid>=1.30
dataclasses>=0.6
typing>=3.7.4
pathlib>=1.0.1
aiohttp>=3.9.1
aiosqlite>=0.19.0
aiologger>=0.7.0
aiofiles>=23.2.1
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-xdist>=3.3.1
